entities | services | generate pdf

+++ src/main/kotlin/br/com/fiap/repository/PedidoRepository.kt
@@ -3,4 +3,4 @@ package br.com.fiap.repository
 import br.com.fiap.entity.Pedido
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface PedidoRepository : JpaRepository<Pedido, Int>
+interface PedidoRepository : JpaRepository<Pedido, Long>
